backend/requirements.txt
# Biometric Attendance System - Backend Requirements
# Core
fastapi==0.109.0
uvicorn==0.27.0
SQLAlchemy==2.0.46
aiomysql==0.3.2
PyMySQL==1.1.2
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1

# Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==46.0.5

# Bio-Recognition (New Stack)
insightface==0.7.3
#onnxruntime-gpu>=1.16.0
onnxruntime>=1.16.0
mediapipe>=0.10.0
opencv-python>=4.9.0

# Image Processing & Utilities
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0
scikit-image>=0.22.0
Pillow>=10.2.0
albumentations>=1.4.0
prettytable>=3.10.0
easydict>=1.13

# Other
python-dotenv==1.2.1
APScheduler==3.11.2
httpx==0.26.0
jinja2==3.1.6
loguru==0.7.2
tqdm==4.67.3
python-dateutil==2.9.0
email-validator==2.3.0
